Update HISTORY and README; add integration tests
Enhanced middleware support and added new sample projects in README.md. Updated HISTORY.md to reflect new features, including mediator statistics endpoints. Added integration tests in UsersControllerTests.cs for various API endpoints to improve test coverage.
